Initial commit of analyze_similarities.py
This script is used for text comparisons between the two question pools.
The Levenshtein similarity, longest common substring and cosine
similarity is computed for each pair of questions from the respective
pools. This is done for the question text alone ("Q") and also for
the concatenation of questions and answers ("QA"). HTML tags are
stripped before comparison.

This can be used to identify identical or similar questions, thus
helping in structuring courses where the novice license is acquired
first, followed by an upgrade class.

Running the program involves many comparisons, thus you cannot expect
this to be quick.