Bug fixed: More than one image per question
In the DLE-2006 question pool, there is a maximum of one image per question. And if there is an image in the question, there are guaranteed to be no images in the answers.

The DLA-2007 question pool is different: it can also contain two images in one question text. And there are questions with one or two images and at the same time the four possible answers also have one image each.

The parser did not previously serve multiple images per question.

convert_to_card2brain.py previously expected either one image per question _or_ images in the  answers (which was only sufficient for  DLE-2006).