new function set_font_size in img-tk
new function set_font_size in img-tk

Images of DLE2006/DLA2007 and the png-converted images from DL-2024 are (on average) of different sizes. Therefore, the font size needs to be adjusted per question pool so that the inserted labels and texts match the size of the displayed electrical components and schematic drawings.