bug #13 fixed (Mix of utf8 and LaTex )
Bug (Mix of utf8 and LaTex in the same term) is fixed.
But it is only solved in the context of an individual analysis. It is only ensured that the individual LaTex term is either completely transformed to utf8  or otherwise left unchanged in LaTex. But still in the same exam question can exist utf8 terms besides LaTex terms. For this  inconsistent appearance I opened issue #20 .